+++ src/engine/Engine.h
@@ -80,7 +80,7 @@ public:
       Attempt to find a feasible solution for the input within a time limit
       (a timeout of 0 means no time limit). Returns true if found, false if infeasible.
     */
-    bool solve( unsigned timeoutInSeconds = 0 );
+    bool solve( double timeoutInSeconds = 0 );
 
     /*
       Minimize the cost function with respect to the current set of linear constraints.
@@ -686,7 +686,7 @@ private:
     /*
       Check whether a timeout value has been provided and exceeded.
     */
-    bool shouldExitDueToTimeout( unsigned timeout ) const;
+    bool shouldExitDueToTimeout( double timeout ) const;
 
     /*
       Evaluate the network on legal inputs; obtain the assignment
@@ -753,7 +753,7 @@ private:
     /*
       Solve the input query with a MILP solver (Gurobi)
     */
-    bool solveWithMILPEncoding( unsigned timeoutInSeconds );
+    bool solveWithMILPEncoding( double timeoutInSeconds );
 
     /*
       Extract the satisfying assignment from the MILP solver
